feat: save/load yaml stats
