+++ organelle_morphology/source.py
@@ -21,6 +21,7 @@ import z5py
 from organelle_morphology.block_mesher import block_mesher
 
 
+from organelle_morphology.statistics import Properties
 from organelle_morphology.util import (
     Cache,
     color_delayed_trimesh_rgba,
@@ -51,7 +52,7 @@ class Data_level:
 
 
 @dataclass
-class SourceMeta:
+class SourceMeta(Properties):
     data_root: Path
     downsampling: list[list[int]]
     levels: list[str]
